Initial refactor of the controller and switch classes to add support for the BFRuntime note: aggregate_controller startup is not complete, data_inspection_t is not getting populated thus breaking the full scenario
