+++ trans_sec/controller/gateway_controller.py
@@ -13,9 +13,15 @@
 from logging import getLogger
 from trans_sec.controller.abstract_controller import AbstractController
 from trans_sec.controller.ddos_sdn_controller import GATEWAY_CTRL_KEY
-from trans_sec.p4runtime_lib.bmv2 import GatewaySwitch
+from trans_sec.p4runtime_lib.gateway_switch import GatewaySwitch as P4RTSwitch
 
 logger = getLogger('gateway_controller')
+
+try:
+    from trans_sec.bfruntime_lib.gateway_switch import (
+        GatewaySwitch as BFRTSwitch)
+except Exception as e:
+    logger.warning('Could not import bfrt classes')
 
 
 class GatewayController(AbstractController):
@@ -28,11 +34,14 @@ class GatewayController(AbstractController):
             platform, p4_build_out, topo, GATEWAY_CTRL_KEY, log_dir, load_p4)
 
     def instantiate_switch(self, sw_info):
-        return GatewaySwitch(
-            p4info_helper=self.p4info_helper,
-            sw_info=sw_info,
-            proto_dump_file='{}/{}-switch-controller.log'.format(
-                self.log_dir, sw_info['name']))
+        if 'arch' in sw_info and sw_info['arch'] == 'tna':
+            return BFRTSwitch(sw_info=sw_info)
+        else:
+            return P4RTSwitch(
+                p4info_helper=self.p4info_helper,
+                sw_info=sw_info,
+                proto_dump_file='{}/{}-switch-controller.log'.format(
+                    self.log_dir, sw_info['name']))
 
     def make_rules(self, sw, north_facing_links, south_facing_links,
                    add_di):
@@ -55,4 +64,5 @@ class GatewayController(AbstractController):
                             ':' + str(device.get('ip_port')))
 
                 if add_di:
-                    sw.add_data_inspection(device['id'], device['mac'])
+                    sw.add_data_inspection(dev_id=device['id'],
+                                           dev_mac=device['mac'])
