+++ trans_sec/p4runtime_lib/p4rt_switch.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019 Cable Television Laboratories, Inc.
+# Copyright (c) 2020 Cable Television Laboratories, Inc.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -10,37 +10,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# Originally copied from:
-#
-# Copyright 2017-present Open Networking Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-import codecs
+
 import logging
 import struct
-from queue import Queue
-from abc import abstractmethod
-from datetime import datetime
+from abc import ABC
 
 import grpc
-from threading import Thread
-
 from p4.tmp import p4config_pb2
 from p4.v1 import p4runtime_pb2, p4runtime_pb2_grpc
 
 from trans_sec.controller.ddos_sdn_controller import GATEWAY_CTRL_KEY
+from trans_sec.switch import SwitchConnection, IterableQueue, GrpcRequestLogger
 from trans_sec.utils.convert import decode_mac
 
 MSG_LOG_MAX_LEN = 1024
@@ -48,48 +28,39 @@ MSG_LOG_MAX_LEN = 1024
 logger = logging.getLogger('switch')
 
 
-class SwitchConnection(object):
+class P4RuntimeSwitch(SwitchConnection, ABC):
 
     def __init__(self, p4info_helper, sw_info, p4_ingress, p4_egress,
                  proto_dump_file=None):
+        super(P4RuntimeSwitch, self).__init__(sw_info)
         self.p4info_helper = p4info_helper
-        self.sw_info = sw_info
         self.p4_ingress = p4_ingress
         self.p4_egress = p4_egress
-        self.name = sw_info['name']
-        self.device_id = sw_info['id']
-        self.grpc_addr = sw_info['grpc']
-        channel = grpc.insecure_channel(self.grpc_addr)
+        grpc_channel = grpc.insecure_channel(self.grpc_addr)
         if proto_dump_file is not None:
             logger.info('Adding interceptor with file - [%s] to device [%s]',
                         proto_dump_file, self.grpc_addr)
             interceptor = GrpcRequestLogger(proto_dump_file)
-            channel = grpc.intercept_channel(channel, interceptor)
+            grpc_channel = grpc.intercept_channel(grpc_channel, interceptor)
 
         logger.info('Creating client stub to channel - [%s] at address - [%s]',
-                    channel, self.grpc_addr)
-        self.client_stub = p4runtime_pb2_grpc.P4RuntimeStub(channel)
+                    grpc_channel, self.grpc_addr)
+        self.client_stub = p4runtime_pb2_grpc.P4RuntimeStub(grpc_channel)
         self.requests_stream = IterableQueue()
-        self.stream_msg_resp = self.client_stub.StreamChannel(iter(
-            self.requests_stream))
+        self.stream_msg_resp = self.client_stub.StreamChannel(
+            iter(self.requests_stream))
         self.proto_dump_file = proto_dump_file
 
-        self.digest_thread = Thread(target=self.receive_digests)
-
     def start_digest_listeners(self):
-        if 'arch' in self.sw_info and self.sw_info['arch'] == 'tofino':
-            logger.info('Tofino currently not supporting digests')
-            pass
-        else:
-            logger.info(
-                'Starting mac_learn_digest Digest for device [%s] named [%s]',
-                self.grpc_addr, self.name)
-            digest_entry, digest_info = self.p4info_helper.build_digest_entry(
-                digest_name="mac_learn_digest")
-            self.write_digest_entry(digest_entry)
+        logger.info(
+            'Starting mac_learn_digest Digest for device [%s] named [%s]',
+            self.grpc_addr, self.name)
+        digest_entry, digest_info = self.p4info_helper.build_digest_entry(
+            digest_name="mac_learn_digest")
+        self.write_digest_entry(digest_entry)
 
-            logger.info('Starting digest threads')
-            self.digest_thread.start()
+        logger.info('Starting digest threads')
+        self.digest_thread.start()
 
     def stop_digest_listeners(self):
         if 'arch' in self.sw_info and self.sw_info['arch'] == 'tofino':
@@ -100,57 +71,8 @@ class SwitchConnection(object):
             self.stream_msg_resp.cancel()
             self.digest_thread.join()
 
-    def receive_digests(self):
-        """
-        Runnable method for self.digest_thread
-        """
-        logger.info("Started listening digest thread on device [%s] with "
-                    "name [%s]", self.grpc_addr, self.name)
-        while True:
-            try:
-                logger.debug('Requesting digests from device [%s]',
-                             self.grpc_addr)
-                digests = self.digest_list()
-                logger.debug('digests from device [%s] - [%s]',
-                             self.grpc_addr, digests)
-                digest_data = digests.digest.data
-                logger.debug('Received digest data from device [%s]: [%s]',
-                             self.grpc_addr, digest_data)
-                self.interpret_digest(digest_data)
-                logger.debug('Interpreted digest data')
-            except Exception as e:
-                logger.error(
-                    'Unexpected error reading digest from device [%s] - [%s]',
-                    self.grpc_addr, e)
-
-    def interpret_digest(self, digest_data):
-        logger.debug("Digest data from switch [%s] - [%s]",
-                     self.name, digest_data)
-
-        if not digest_data or len(digest_data) == 0:
-            logger.warning('No digest data to process')
-            return
-        for members in digest_data:
-            logger.debug("Digest members: %s", members)
-            if members.WhichOneof('data') == 'struct':
-                source_mac = decode_mac(members.struct.members[0].bitstring)
-                if source_mac:
-                    logger.debug('Digest MAC Address is: %s', source_mac)
-                    ingress_port = int(
-                        codecs.encode(members.struct.members[1].bitstring,
-                                      'hex'), 16)
-                    logger.debug('Digest Ingress Port is %s', ingress_port)
-                    self.add_data_forward(source_mac, ingress_port)
-                else:
-                    logger.warning('Could not retrieve source_mac from digest')
-            else:
-                logger.warning('Digest could not be processed - [%s]',
-                               digest_data)
-
-        logger.info('Completed digest processing')
-
-    def __get_table_entry(self, table_name, action_name, match_fields,
-                          action_params, ingress_class=True):
+    def get_table_entry(self, table_name, action_name, match_fields,
+                        action_params, ingress_class=True):
         if ingress_class:
             tbl_class = self.p4_ingress
         else:
@@ -162,26 +84,11 @@ class SwitchConnection(object):
             action_name='{}.{}'.format(tbl_class, action_name),
             action_params=action_params)
 
-    def insert_p4_table_entry(self, table_name, action_name, match_fields,
-                              action_params, ingress_class=True,
-                              election_high=0, election_low=1):
-
-        table_entry = self.__get_table_entry(
-            table_name, action_name, match_fields, action_params,
-            ingress_class)
-        logger.debug(
-            'Writing table entry to device [%s] table [%s], '
-            'with action name - [%s], '
-            'match fields - [%s], action_params - [%s]',
-            self.grpc_addr, table_name, action_name, match_fields,
-            action_params)
-        self.write_table_entry(table_entry, election_high, election_low)
-
     def delete_p4_table_entry(self, table_name, action_name, match_fields,
                               action_params=None, ingress_class=True,
                               election_high=0, election_low=1):
 
-        table_entry = self.__get_table_entry(
+        table_entry = self.get_table_entry(
             table_name, action_name, match_fields, action_params,
             ingress_class)
         logger.debug(
@@ -579,47 +486,3 @@ class SwitchConnection(object):
         request.election_id.high = high
         request.election_id.low = low
         return request
-
-
-class GrpcRequestLogger(grpc.UnaryUnaryClientInterceptor,
-                        grpc.UnaryStreamClientInterceptor):
-    """Implementation of a gRPC interceptor that logs request to a file"""
-
-    def __init__(self, log_file):
-        self.log_file = log_file
-        with open(self.log_file, 'w') as f:
-            # Clear content if it exists.
-            f.write("")
-
-    def log_message(self, method_name, body):
-        with open(self.log_file, 'a') as f:
-            ts = datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S.%f')[:-3]
-            msg = str(body)
-            f.write("\n[%s] %s\n---\n" % (ts, method_name))
-            if len(msg) < MSG_LOG_MAX_LEN:
-                f.write(str(body))
-            else:
-                f.write(
-                    "Message too long (%d bytes)! Skipping log...\n" % len(
-                        msg))
-            f.write('---\n')
-
-    def intercept_unary_unary(self, continuation, client_call_details,
-                              request):
-        self.log_message(client_call_details.method, request)
-        return continuation(client_call_details, request)
-
-    def intercept_unary_stream(self, continuation, client_call_details,
-                               request):
-        self.log_message(client_call_details.method, request)
-        return continuation(client_call_details, request)
-
-
-class IterableQueue(Queue):
-    _sentinel = object()
-
-    def __iter__(self):
-        return iter(self.get, self._sentinel)
-
-    def close(self):
-        self.put(self._sentinel)
